Reorganize metrics code and refactor a global, per review by Devon
