+++ app/javascript/lib/scp-api.js
@@ -10,7 +10,9 @@ import camelcaseKeys from 'camelcase-keys'
 import _compact from 'lodash/compact'
 
 import { accessToken } from './../components/UserProvider'
-import { logFilterSearch, logSearch, logDownloadAuthorization } from './metrics-api'
+import {
+  logFilterSearch, logSearch, logDownloadAuthorization, mapFiltersForLogging
+} from './scp-api-metrics'
 
 // If true, returns mock data for all API responses.  Only for dev.
 let globalMock = false
@@ -31,31 +33,6 @@ if (
   defaultInit.headers = Object.assign(defaultInit.headers, {
     'Authorization': `Bearer ${accessToken}`
   })
-}
-
-/**
- * Populates global id-to-name map of retrieved filters, for easier analytics.
- * See downstream use of window.SCP.filterNamesById in metrics-api.js.
- */
-function mapFilters(facetsOrFilters, isFacets=false) {
-  // If testing, skip.  Tech debt to reconsider later.
-  if ('SCP' in window === false) return
-
-  // This construct is kludgy, but helps cohesion and encapsulation
-  // by putting related dense code here instead of in the calling functions
-  if (isFacets) {
-    const facets = facetsOrFilters
-    facets.map(facet => {
-      facet.filters.map(filter => {
-        window.SCP.filterNamesById[filter.id] = filter.name
-      })
-    })
-  } else {
-    const filters = facetsOrFilters
-    filters.map(filter => {
-      window.SCP.filterNamesById[filter.id] = filter.name
-    })
-  }
 }
 
 /**
@@ -96,7 +73,7 @@ export async function fetchAuthCode(mock=false) {
 export async function fetchFacets(mock=false) {
   const facets = await scpApi('/search/facets', defaultInit, mock)
 
-  mapFilters(facets, true)
+  mapFiltersForLogging(facets, true)
 
   return facets
 }
@@ -159,7 +136,7 @@ export async function fetchFacetFilters(facet, query, mock=false) {
 
   const filters = await scpApi(pathAndQueryString, defaultInit, mock)
 
-  mapFilters(filters)
+  mapFiltersForLogging(filters)
 
   return filters
 }
