+++ app/views/layouts/application.html.erb
@@ -37,10 +37,6 @@
       }
       window.SCP.userId = Cookies.get('user_id');
 
-      // Maps filter IDs to filter names for easier analytics.
-      // See use in scp-api.js and metrics-api.js.
-      window.SCP.filterNamesById = {};
-
   </script>
 
   <script type="text/javascript"  nonce="<%= content_security_policy_script_nonce %>">
